+++ test/git_hub_actions/versions_test.exs
@@ -19,7 +19,7 @@ defmodule GitHubActions.VersionsTest do
 
   test "from_config/0" do
     assert List.last(Versions.from_config()) ==
-             [otp: ["27.0/2"], elixir: ["1.17.0/3", "1.18.0"]]
+             [otp: ["27.0/2"], elixir: ["1.17.0/3", "1.18.0/1"]]
   end
 
   describe "get/2" do
@@ -135,8 +135,9 @@ defmodule GitHubActions.VersionsTest do
                "1.17.1",
                "1.17.2",
                "1.17.3",
-               # v1.18.0
-               "1.18.0"
+               # v1.18.0/1
+               "1.18.0",
+               "1.18.1"
              ]
     end
 
@@ -290,7 +291,8 @@ defmodule GitHubActions.VersionsTest do
                %Version{major: 1, minor: 17, patch: 1},
                %Version{major: 1, minor: 17, patch: 2},
                %Version{major: 1, minor: 17, patch: 3},
-               %Version{major: 1, minor: 18, patch: 0}
+               %Version{major: 1, minor: 18, patch: 0},
+               %Version{major: 1, minor: 18, patch: 1}
              ]
     end
 
@@ -328,7 +330,7 @@ defmodule GitHubActions.VersionsTest do
                %Version{major: 1, minor: 15, patch: 8},
                %Version{major: 1, minor: 16, patch: 3},
                %Version{major: 1, minor: 17, patch: 3},
-               %Version{major: 1, minor: 18, patch: 0}
+               %Version{major: 1, minor: 18, patch: 1}
              ]
     end
   end
@@ -368,9 +370,9 @@ defmodule GitHubActions.VersionsTest do
         %Version{major: 23, minor: 3}
       ]
 
-      assert Versions.compatible(@versions, :otp, elixir: "1.11") == otp
+      assert Versions.compatible_to(@versions, :otp, elixir: "1.11") == otp
 
-      assert Versions.compatible(@versions, :otp, elixir: "1.11.4") ==
+      assert Versions.compatible_to(@versions, :otp, elixir: "1.11.4") ==
                otp ++
                  [
                    %Version{major: 24, minor: 0},
@@ -381,7 +383,7 @@ defmodule GitHubActions.VersionsTest do
     end
 
     test "otp versions for multiple elixir versions" do
-      assert Versions.compatible(@versions, :otp, elixir: "1.11.0/4") == [
+      assert Versions.compatible_to(@versions, :otp, elixir: "1.11.0/4") == [
                %Version{major: 21, minor: 0},
                %Version{major: 21, minor: 1},
                %Version{major: 21, minor: 2},
@@ -402,18 +404,18 @@ defmodule GitHubActions.VersionsTest do
     end
 
     test "raises error for invalid versions" do
-      message = "compatible/3 expected a table of versions as first argument, got: [:a]"
+      message = "compatible_to/3 expected a table of versions as first argument, got: [:a]"
 
       assert_raise ArgumentError, message, fn ->
-        Versions.compatible([:a], :a, b: "1")
+        Versions.compatible_to([:a], :a, b: "1")
       end
     end
 
     test "raises error for invalid version" do
-      message = "compatible/3 expected a list of versions for :elixir, got: [:a]"
+      message = "compatible_to/3 expected a list of versions for :elixir, got: [:a]"
 
       assert_raise ArgumentError, message, fn ->
-        Versions.compatible(@versions, :otp, elixir: [:a])
+        Versions.compatible_to(@versions, :otp, elixir: [:a])
       end
     end
   end
@@ -438,7 +440,7 @@ defmodule GitHubActions.VersionsTest do
         |> Versions.filter("~> 1.12")
         |> Versions.latest_minor()
 
-      otp = @versions |> Versions.compatible(:otp, elixir: elixir) |> Versions.latest_major()
+      otp = @versions |> Versions.compatible_to(:otp, elixir: elixir) |> Versions.latest_major()
 
       assert Versions.incompatible(@versions, otp: otp, elixir: elixir) == [
                [
@@ -459,7 +461,7 @@ defmodule GitHubActions.VersionsTest do
                ],
                [
                  otp: %Version{major: 22, minor: 3},
-                 elixir: %Version{major: 1, minor: 18, patch: 0}
+                 elixir: %Version{major: 1, minor: 18, patch: 1}
                ],
                [
                  otp: %Version{major: 23, minor: 3},
@@ -475,7 +477,7 @@ defmodule GitHubActions.VersionsTest do
                ],
                [
                  otp: %Version{major: 23, minor: 3},
-                 elixir: %Version{major: 1, minor: 18, patch: 0}
+                 elixir: %Version{major: 1, minor: 18, patch: 1}
                ],
                [
                  otp: %Version{major: 24, minor: 3},
@@ -483,7 +485,7 @@ defmodule GitHubActions.VersionsTest do
                ],
                [
                  otp: %Version{major: 24, minor: 3},
-                 elixir: %Version{major: 1, minor: 18, patch: 0}
+                 elixir: %Version{major: 1, minor: 18, patch: 1}
                ],
                [
                  otp: %Version{major: 25, minor: 3},
@@ -565,7 +567,7 @@ defmodule GitHubActions.VersionsTest do
   end
 
   describe "matrix/2" do
-    test "create matrix for elixir requirement '> 1.10'" do
+    test "creates matrix with default mode :exclude" do
       assert Versions.matrix(elixir: ">= 1.10.0 and < 1.13.3", otp: ">= 20.0.0 and < 24.0.0") == [
                elixir: [
                  Version.parse!("1.10.4"),
@@ -585,10 +587,42 @@ defmodule GitHubActions.VersionsTest do
              ]
     end
 
+    test "creates matrix with mode :include" do
+      assert Versions.matrix(
+               elixir: ">= 1.10.0 and < 1.13.3",
+               otp: ">= 20.0.0 and < 24.0.0",
+               mode: :include
+             ) == [
+               include: [
+                 [elixir: Version.parse!("1.10.4"), otp: Version.parse!("21.3")],
+                 [elixir: Version.parse!("1.10.4"), otp: Version.parse!("22.3")],
+                 [elixir: Version.parse!("1.10.4"), otp: Version.parse!("23.3")],
+                 [elixir: Version.parse!("1.11.4"), otp: Version.parse!("21.3")],
+                 [elixir: Version.parse!("1.11.4"), otp: Version.parse!("22.3")],
+                 [elixir: Version.parse!("1.11.4"), otp: Version.parse!("23.3")],
+                 [elixir: Version.parse!("1.12.3"), otp: Version.parse!("22.3")],
+                 [elixir: Version.parse!("1.12.3"), otp: Version.parse!("23.3")],
+                 [elixir: Version.parse!("1.13.2"), otp: Version.parse!("22.3")],
+                 [elixir: Version.parse!("1.13.2"), otp: Version.parse!("23.3")]
+               ]
+             ]
+    end
+
     test "create matrix for elixir requirement '== 1.13.3'" do
       assert Versions.matrix(elixir: "== 1.13.3", otp: "== 23.0.0") == [
                elixir: [Version.parse!("1.13.3")],
                otp: [Version.parse!("23.0")]
+             ]
+    end
+
+    test "create matrix for elixir requirement '== 1.13.3' and mode: :include" do
+      assert Versions.matrix(elixir: "== 1.13.3", otp: "== 23.0.0", mode: :include) == [
+               include: [
+                 [
+                   elixir: Version.parse!("1.13.3"),
+                   otp: Version.parse!("23.0")
+                 ]
+               ]
              ]
     end
   end